Remove some unnecessary python functional wrappers
See #61544 for an example of issues created by functional wrappers. In this
case, these are directly wrapping the native function with no added
functionality. One exception was `bilinear` which was just missing the default
argument in C++, but was otherwise the same.

I've kept the symbol `torch.functional.istft` because it looks like public API,
but it could just as easily be moved to `_torch_docs.py`.

ghstack-source-id: 3a920a2e52ec3605f5b1c7a0c890a9e4f7be7b93
Pull Request resolved: https://github.com/pytorch/pytorch/pull/61608